Rename additional designs
