+++ src/components/Design2/MainContent.component.jsx
